backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==25.12.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Django==6.0
django-debug-toolbar==6.1.0
django-environ==0.12.0
django-filter==25.2
django-picklefield==3.4.0
django-q2==1.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
execnet==2.1.2
flake8==7.3.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
isort==7.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pytest==9.0.2
pytest-django==4.11.1
pytest-xdist==3.8.0
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
wheel==0.45.1
